feat(lib): update main library implementations with formatting adjustments
The original template that I based this one off of is a very good looking
template, but my text seemed to be a bit longer that what it afforded. Also
needed to make some localization specific changes, e.g. redaction of phone
number and removal of the inclusion of the profile picture directly on the
resume (generally seen as a no-go for US hiring)